Tests: coverage: Disabled coverage reporting for 2 models.py files, due to incorrect coverage reporting of model code (see django-nose Issue #108).
Details: Prior to this change, the 2 files were showing in the coverage reports as partially covered, with some lines that are covered by tests shown as uncovered/missing. With this change, both files are completely excluded from the coverage metrics, which is still undesirable but is deemed preferable (at least for now) to having "false positive" missing line coverage metrics.